同步Connect和Send

+++ window/NetProtocol/NetProtocol/SockPool.h
@@ -37,7 +37,7 @@ typedef struct Manager
 struct transstruct {
 	unsigned short  dstport;   //目的端口号
 	unsigned short  srcport;   //源端口号
-	unsigned int  function;     //绑定端口号
+	unsigned short  function;     //绑定端口号
 	int    datalength;             //数据长度
 	char srcip[20];                //原地址ip
 	char dstip[20];                //目标地址ip
